+++ src/routes/categories.js
@@ -1,4 +1,5 @@
 const { Category, validate } = require('../models/category')
+const { ObjectId } = require('mongodb')
 const express = require('express')
 const router = express.Router()
 
@@ -9,6 +10,8 @@ router.get('/', async (req, res) => {
 })
 
 router.get('/:id', async (req, res) => {
+    const isValidId = ObjectId.isValid(req.params.id)
+    if (!isValidId) return res.status(400).send('Id is invalid')
     const category = await Category.findById(req.params.id)
     if (!category) return res.status(404).send('category not exist')
     res.send(category);
@@ -25,12 +28,16 @@ router.post('/', async (req, res) => {
 router.put('/:id', async (req, res) => {
     const { error } = validate(req.body)
     if (error) return res.status(400).send(errors.details[0].message)
+    const isValidId = ObjectId.isValid(req.params.id)
+    if (!isValidId) return res.status(400).send('Id is invalid')
     const category = Category.findByIdAndUpdate(req.params.id, { name: req.body.name }, { new: true })
     if (!category) return res.status(404).send('category with given id not exist')
     res.send(category)
 })
 
 router.delete('/:id', async (req, res) => {
+    const isValidId = ObjectId.isValid(req.params.id)
+    if (!isValidId) return res.status(400).send('Id is invalid')
     const category = await Category.findByIdAndRemove(req.params.id)
     if (!category) return res.status(404).send('category with given id not exist')
     res.send(category)
